Finalização do C4S2

+++ src/modules/accounts/useCases/updateUserAvatar/UpdateUserAvatarUseCase.ts
@@ -1,10 +1,10 @@
 import { inject, injectable } from 'tsyringe';
 
-import { ICreateUserDTO } from '../../dtos/ICreateUserDTO';
-import { IUsersRepository } from "../../repositories/IUsersRepository";
+import { IUsersRepository } from "@modules/accounts/repositories/IUsersRepository";
 
-import { deleteFile } from '../../../../utils/file';
-import AppError from '../../../../errors/AppError';
+import AppError from '@shared/errors/AppError';
+
+import { deleteFile } from '@utils/file';
 
 interface IRequest {
     user_id: string;
@@ -25,7 +25,9 @@ class UpdateUserAvatarUseCase {
 
         const user = await this.usersRepository.findById(user_id);
 
-        await deleteFile(`./tmp/avatar/${user.avatar}`);
+        if (user.avatar) {
+            await deleteFile(`./tmp/avatar/${user.avatar}`);
+        }
 
         user.avatar = avatar_file;
